% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_my_content.R
\name{add_my_content}
\alias{add_my_content}
\title{Helper to create/update a post fork}
\usage{
add_my_content(post_dir, roweb2_clone_path, slug, date,
  message = "work on my post", push = FALSE)
}
\arguments{
\item{post_dir}{Directory where your post lives}

\item{roweb2_clone_path}{Directory that's your local roweb2 clone}

\item{slug}{Slug of your post, it should correspond to the slug
in the post filename.}

\item{date}{Date of your post, it should correspond to the date
in the post filename.}

\item{message}{Commit message. Fine if not very informative,
since we'll squash and merge.}

\item{push}{Whether to push. See Details.}
}
\description{
Helper to create or update the branch of
 your roweb2 fork containing your blog post/tech note,
 from which you'll make a PR to ropensci/roweb2.
}
\details{
Only use \code{push=TRUE} if you know you can push using \code{git2r}.
How does one know? Well if you've been able to use \code{usethis::use_github}.
At the moment, chose \code{FALSE} if you use Windows!
}
\examples{
\dontrun{
post_dir <- "C:/Users/Maelle/Documents/ropensci/orcid_note"
slug <- "orcid"
date <- "2018-10-08"
roweb2_clone_path <- "C:/Users/Maelle/Documents/ropensci/roweb2"
add_my_content(post_dir = post_dir,
            roweb2_clone_path = roweb2_clone_path,
            slug = slug, date = date,
           message = "work on my post")
}
}
