% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_newsletter_content.R
\name{create_newsletter_content}
\alias{create_newsletter_content}
\title{Create newsletter content for Sendgrid}
\usage{
create_newsletter_content()
}
\value{
Text copied to clipboard to be pasted into Sendgrid
}
\description{
Create newsletter content for Sendgrid
}
