% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_post_project.R
\name{create_post_project}
\alias{create_post_project}
\title{Create a RStudio project with a blog post template}
\usage{
create_post_project(dir = tempdir(), slug, date, type = "Rmd",
  category = "blog")
}
\arguments{
\item{dir}{where to create the project}

\item{slug}{Slug describing the blog post, e.g
"thanking-reviewers-in-metadata". Use hyphens, not space.}

\item{date}{yyyy-mm-dd}

\item{type}{Either 'Rmd' or 'md'}

\item{category}{Either 'blog' or 'technotes'}
}
\value{
It opens a new project.
}
\description{
Create a RStudio project with a blog post template
}
\examples{
\dontrun{
create_post_project(tempdir(),
                    "test", "2018-10-15",
                    type = "Rmd",
                    category = "blog")
create_post_project(tempdir(),
                    "test2", "2018-10-15",
                    type = "md",
                    category = "technotes")
}
}
