% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint-posts.R
\name{ro_lint_md}
\alias{ro_lint_md}
\title{Lint Markdown post for rOpenSci blog}
\usage{
ro_lint_md(path)
}
\arguments{
\item{path}{Path to the Markdown post (not source Rmd!)}
}
\description{
Lint Markdown post for rOpenSci blog
}
\examples{
\dontrun{
path <- system.file(file.path("examples", "bad-no-alt.md"),
                    package = "roblog")
}
}
