% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ro.R
\name{theme_ro}
\alias{theme_ro}
\title{rOpenSci ggplot2 theme}
\usage{
theme_ro(base_size = 12, axis_title_size = 12, axis_text_size = 12, ...)
}
\arguments{
\item{base_size}{size parameter}

\item{axis_title_size}{size parameter}

\item{...}{other parameters passed to \code{hrbrthemes::theme_ipsum}}
}
\value{
ggplot2 theme
}
\description{
rOpenSci ggplot2 theme, which is just a wrapper
 around Bob Rudis' hrbrthemes.
}
\examples{
\dontrun{
library("ggplot2")
df <- data.frame(
gp = factor(rep(letters[1:3], each = 10)),
y = rnorm(30)
)
ggplot(df, aes(gp, y)) +
  geom_point() +
  ggtitle("wow title LHS great",
          subtitle = "nice plot") +
  theme_ro()
  }
}
