% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_blog_post.R
\name{ro_blog_post_rmd}
\alias{ro_blog_post_rmd}
\alias{ro_blog_post_md}
\title{Create a new rOpenSci (R) Markdown Blog Post}
\usage{
ro_blog_post_rmd()

ro_blog_post_md()
}
\description{
Create a new rOpenSci Blog Post,
either R Markdown (`ro_blog_post_rmd()`) or Markdown (`ro_blog_post_md()`)
}
\details{
Call it via the add-in or directly.
}
