% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_blog_post.R
\name{ro_blog_post_author}
\alias{ro_blog_post_author}
\title{Create a new rOpenSci author file}
\usage{
ro_blog_post_author()
}
\description{
Create a new author file, in RStudio.
}
\details{
Call it via the add-in or directly or get the \href{https://ropensci-org.github.io/blog-guidance/templatemd.html}{templates online}.

In any case, an internet connection is required as templates are downloaded
fresh from their source

\code{ro_blog_post_author()} creates Markdown files,
RStudio might warn you against saving them as ".md" but ignore that.
}
